refactory: 코어, 유틸 분류
