+++ lib/core/ai_route_helper.dart
