+++ lib/core/bottom_bar_provider.dart
