+++ lib/views/my_page/plan_list_page.dart
@@ -2,8 +2,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:travel_muse_app/constants/app_text_styles.dart';
-import 'package:travel_muse_app/core/format_month_day.dart';
-import 'package:travel_muse_app/core/format_region.dart';
+import 'package:travel_muse_app/utills/format_month_day.dart';
+import 'package:travel_muse_app/utills/format_region.dart';
 import 'package:travel_muse_app/core/widgets/bottom_bar.dart';
 import 'package:travel_muse_app/providers/plan/schedule/schedule_provider.dart';
 import 'package:travel_muse_app/viewmodels/user/profile_view_model.dart';
